Remove the parsing of more from the zwsdecoder.c
Currently the more information is parsed in the zwsdecoder. This means our decoder would only work in case we have "normal" ZeroMQ frames. This code is the first step to give zwssock support for permessage-deflate as part of https://tools.ietf.org/html/draft-ietf-hybi-permessage-compression-28

I would strongly urge not to blindly merge this commit, until properly reviewed. I haved used valgrind on the code, and did several tests with browser interaction. The step by step commits are compiled from the big permessage-deflate implementation to show the steps towards it.